refactor(test_LSFIR): restructure test_LSFIR into several modules to improve splitting
