+++ test/test_lsfir/test_LSFIR.py
@@ -1,197 +1,25 @@
 import os
 import pathlib
-from typing import Callable, cast, Optional, Union
+from typing import cast
 
 import hypothesis.strategies as hst
 import numpy as np
 import pytest
 import scipy.signal as dsp
 from hypothesis import given, HealthCheck, settings
-from hypothesis.strategies import composite
 from matplotlib import pyplot as plt
-from numpy.testing import assert_allclose, assert_almost_equal
-
-from PyDynamic.misc.filterstuff import kaiser_lowpass
-from PyDynamic.misc.SecondOrderSystem import sos_phys2filter
-from PyDynamic.misc.tools import (
-    complex_2_real_imag,
-    make_semiposdef,
-    real_imag_2_complex,
-)
+from numpy.testing import assert_allclose
 
 # noinspection PyProtectedMember
+from PyDynamic import FIRuncFilter, kaiser_lowpass
 from PyDynamic.model_estimation.fit_filter import (
     invLSFIR,
-    invLSFIR_unc,
-    invLSFIR_uncMC,
     LSFIR,
 )
-from PyDynamic.uncertainty.propagate_filter import FIRuncFilter
-from .conftest import (
-    _print_during_test_to_avoid_timeout,
+from .conftest import weights
+from ..conftest import (
     hypothesis_dimension,
-    hypothesis_float_vector,
-    scale_matrix_or_vector_to_convex_combination,
 )
-
-
-@pytest.fixture(scope="module")
-def random_number_generator():
-    return np.random.default_rng(1)
-
-
-@pytest.fixture(scope="module")
-def measurement_system():
-    f0 = 36e3
-    S0 = 0.4
-    delta = 0.01
-    return {"f0": f0, "S0": S0, "delta": delta}
-
-
-@composite
-def weights(
-    draw: Callable, guarantee_vector: Optional[bool] = False
-) -> Union[np.ndarray, None]:
-    valid_vector_strategy = hypothesis_float_vector(
-        length=400, min_value=0, max_value=1, exclude_min=True
-    )
-    valid_weight_strategies = (
-        valid_vector_strategy
-        if guarantee_vector
-        else (valid_vector_strategy, hst.just(None))
-    )
-    unscaled_weights = draw(hst.one_of(valid_weight_strategies))
-    if unscaled_weights is not None:
-        return scale_matrix_or_vector_to_convex_combination(unscaled_weights)
-
-
-@pytest.fixture(scope="module")
-def sampling_freq():
-    return 500e3
-
-
-@pytest.fixture(scope="module")
-def freqs():
-    return np.linspace(0, 120e3, 200)
-
-
-@pytest.fixture(scope="module")
-def monte_carlo(
-    measurement_system, random_number_generator, sampling_freq, freqs, complex_freq_resp
-):
-    udelta = 0.1 * measurement_system["delta"]
-    uS0 = 0.001 * measurement_system["S0"]
-    uf0 = 0.01 * measurement_system["f0"]
-
-    runs = 10000
-    MCS0 = random_number_generator.normal(
-        loc=measurement_system["S0"], scale=uS0, size=runs
-    )
-    MCd = random_number_generator.normal(
-        loc=measurement_system["delta"], scale=udelta, size=runs
-    )
-    MCf0 = random_number_generator.normal(
-        loc=measurement_system["f0"], scale=uf0, size=runs
-    )
-    HMC = np.empty((runs, len(freqs)), dtype=complex)
-    for index, mcs0_mcd_mcf0 in enumerate(zip(MCS0, MCd, MCf0)):
-        bc_, ac_ = sos_phys2filter(mcs0_mcd_mcf0[0], mcs0_mcd_mcf0[1], mcs0_mcd_mcf0[2])
-        b_, a_ = dsp.bilinear(bc_, ac_, sampling_freq)
-        HMC[index, :] = dsp.freqz(b_, a_, 2 * np.pi * freqs / sampling_freq)[1]
-
-    H = complex_2_real_imag(complex_freq_resp)
-    assert_allclose(
-        H,
-        np.load(
-            os.path.join(
-                pathlib.Path(__file__).parent.resolve(),
-                "reference_arrays",
-                "test_LSFIR_H.npz",
-            ),
-        )["H"],
-    )
-    uAbs = np.std(np.abs(HMC), axis=0)
-    assert_allclose(
-        uAbs,
-        np.load(
-            os.path.join(
-                pathlib.Path(__file__).parent.resolve(),
-                "reference_arrays",
-                "test_LSFIR_uAbs.npz",
-            ),
-        )["uAbs"],
-        rtol=3.5e-2,
-    )
-    uPhas = np.std(np.angle(HMC), axis=0)
-    assert_allclose(
-        uPhas,
-        np.load(
-            os.path.join(
-                pathlib.Path(__file__).parent.resolve(),
-                "reference_arrays",
-                "test_LSFIR_uPhas.npz",
-            ),
-        )["uPhas"],
-        rtol=4.3e-2,
-    )
-    UH = np.cov(np.hstack((np.real(HMC), np.imag(HMC))), rowvar=False)
-    UH = make_semiposdef(UH)
-    assert_allclose(
-        UH,
-        np.load(
-            os.path.join(
-                pathlib.Path(__file__).parent.resolve(),
-                "reference_arrays",
-                "test_LSFIR_UH.npz",
-            ),
-        )["UH"],
-        atol=1,
-    )
-    return {"H": H, "uAbs": uAbs, "uPhas": uPhas, "UH": UH}
-
-
-@pytest.fixture(scope="module")
-def complex_H_with_UH(monte_carlo):
-    return {
-        "H": real_imag_2_complex(monte_carlo["H"]),
-        "UH": monte_carlo["UH"],
-    }
-
-
-@pytest.fixture(scope="module")
-def digital_filter(measurement_system, sampling_freq):
-    # transform continuous system to digital filter
-    bc, ac = sos_phys2filter(
-        measurement_system["S0"], measurement_system["delta"], measurement_system["f0"]
-    )
-    assert_almost_equal(bc, [20465611686.098896])
-    assert_allclose(ac, np.array([1.00000000e00, 4.52389342e03, 5.11640292e10]))
-    b, a = dsp.bilinear(bc, ac, sampling_freq)
-    assert_allclose(
-        b, np.array([0.019386043211510096, 0.03877208642302019, 0.019386043211510096])
-    )
-    assert_allclose(a, np.array([1.0, -1.7975690550957188, 0.9914294872108197]))
-    return {"b": b, "a": a}
-
-
-@pytest.fixture(scope="module")
-def complex_freq_resp(measurement_system, sampling_freq, freqs, digital_filter):
-    Hf = dsp.freqz(
-        digital_filter["b"],
-        digital_filter["a"],
-        2 * np.pi * freqs / sampling_freq,
-    )[1]
-    assert_allclose(
-        Hf,
-        np.load(
-            os.path.join(
-                pathlib.Path(__file__).parent.resolve(),
-                "reference_arrays",
-                "test_LSFIR_Hf.npz",
-            ),
-        )["Hf"],
-    )
-    return Hf
 
 
 @pytest.fixture(scope="module")
@@ -427,7 +255,7 @@ def test_digital_deconvolution_FIR_example_figure_5(
 
 
 def test_digital_deconvolution_FIR_example_figure_6(
-    simulated_measurement_input_and_output, LSFIR_filter_fit, fir_unc_filter
+    simulated_measurement_input_and_output, fir_unc_filter
 ):
     plt.figure(figsize=(16, 8))
     plt.plot(
@@ -467,196 +295,6 @@ def test_digital_deconvolution_FIR_example_figure_7(
     plt.subplots_adjust(left=0.15, right=0.95)
     plt.tick_params(which="both", labelsize=16)
     plt.xlim(1.9, 2.4)
-
-
-@given(hypothesis_dimension(min_value=4, max_value=8))
-@settings(
-    deadline=None,
-    suppress_health_check=[
-        *settings.default.suppress_health_check,
-        HealthCheck.too_slow,
-        HealthCheck.function_scoped_fixture,
-    ],
-)
-@pytest.mark.slow
-def test_compare_LSFIR_with_zero_to_None_uncertainties_with_svd_for_fitting_one_over_H(
-    capsys, monte_carlo, freqs, sampling_freq, filter_order
-):
-    b_fir_svd = LSFIR(
-        H=monte_carlo["H"],
-        UH=np.zeros_like(monte_carlo["UH"]),
-        N=filter_order,
-        tau=filter_order // 2,
-        f=freqs,
-        Fs=sampling_freq,
-        inv=True,
-    )[0]
-    b_fir_none = LSFIR(
-        H=monte_carlo["H"],
-        N=filter_order,
-        tau=filter_order // 2,
-        f=freqs,
-        Fs=sampling_freq,
-        inv=True,
-        UH=None,
-    )[0]
-    _print_during_test_to_avoid_timeout(capsys)
-    assert_allclose(b_fir_svd, b_fir_none)
-
-
-@given(hypothesis_dimension(min_value=4, max_value=8))
-@settings(
-    deadline=None,
-    suppress_health_check=[
-        *settings.default.suppress_health_check,
-        HealthCheck.too_slow,
-        HealthCheck.function_scoped_fixture,
-    ],
-)
-@pytest.mark.slow
-def test_compare_LSFIR_with_zero_to_None_uncertainties_and_mc_for_fitting_one_over_H(
-    capsys, monte_carlo, freqs, sampling_freq, filter_order
-):
-    b_fir_mc = LSFIR(
-        H=monte_carlo["H"],
-        UH=np.zeros_like(monte_carlo["UH"]),
-        N=filter_order,
-        tau=filter_order // 2,
-        f=freqs,
-        Fs=sampling_freq,
-        inv=True,
-        mc_runs=2,
-    )[0]
-    b_fir_none = LSFIR(
-        H=monte_carlo["H"],
-        N=filter_order,
-        tau=filter_order // 2,
-        f=freqs,
-        Fs=sampling_freq,
-        inv=True,
-        UH=None,
-    )[0]
-    _print_during_test_to_avoid_timeout(capsys)
-    assert_allclose(b_fir_mc, b_fir_none)
-
-
-@given(hypothesis_dimension(min_value=4, max_value=8))
-@settings(
-    deadline=None,
-    suppress_health_check=[
-        *settings.default.suppress_health_check,
-        HealthCheck.too_slow,
-        HealthCheck.function_scoped_fixture,
-    ],
-)
-@pytest.mark.slow
-def test_compare_LSFIR_with_zero_to_None_uncertainties_and_mc_for_fitting_H_directly(
-    capsys, monte_carlo, freqs, sampling_freq, filter_order
-):
-    b_fir_mc = LSFIR(
-        H=monte_carlo["H"],
-        UH=np.zeros_like(monte_carlo["UH"]),
-        N=filter_order,
-        tau=filter_order // 2,
-        f=freqs,
-        Fs=sampling_freq,
-        inv=False,
-        mc_runs=2,
-    )[0]
-    b_fir_none = LSFIR(
-        H=monte_carlo["H"],
-        N=filter_order,
-        tau=filter_order // 2,
-        f=freqs,
-        Fs=sampling_freq,
-        inv=False,
-        UH=None,
-    )[0]
-    _print_during_test_to_avoid_timeout(capsys)
-    assert_allclose(b_fir_mc, b_fir_none)
-
-
-@given(hypothesis_dimension(min_value=4, max_value=8), weights())
-@settings(
-    deadline=None,
-    suppress_health_check=[
-        *settings.default.suppress_health_check,
-        HealthCheck.too_slow,
-        HealthCheck.function_scoped_fixture,
-    ],
-    max_examples=10,
-)
-@pytest.mark.slow
-def test_usual_call_LSFIR_for_fitting_H_directly_with_svd(
-    capsys, monte_carlo, freqs, sampling_freq, filter_order, weight_vector
-):
-    LSFIR(
-        H=monte_carlo["H"],
-        N=filter_order,
-        f=freqs,
-        Fs=sampling_freq,
-        tau=filter_order // 2,
-        weights=weight_vector,
-        inv=True,
-        UH=monte_carlo["UH"],
-    )
-    _print_during_test_to_avoid_timeout(capsys)
-
-
-@given(hypothesis_dimension(min_value=4, max_value=8), hst.booleans())
-@settings(
-    deadline=None,
-    suppress_health_check=[
-        *settings.default.suppress_health_check,
-        HealthCheck.too_slow,
-    ],
-    max_examples=10,
-)
-def test_usual_call_LSFIR_with_None_uncertainties(
-    monte_carlo, freqs, sampling_freq, filter_order, fit_reciprocal
-):
-    LSFIR(
-        H=monte_carlo["H"],
-        N=filter_order,
-        f=freqs,
-        Fs=sampling_freq,
-        tau=filter_order // 2,
-        inv=fit_reciprocal,
-        UH=None,
-    )
-
-
-@given(
-    hypothesis_dimension(min_value=4, max_value=8),
-    weights(),
-    hst.booleans(),
-    hst.booleans(),
-)
-@settings(
-    deadline=None,
-    suppress_health_check=[
-        *settings.default.suppress_health_check,
-        HealthCheck.function_scoped_fixture,
-    ],
-    max_examples=10,
-)
-@pytest.mark.slow
-def test_usual_call_LSFIR_with_mc(
-    capsys, monte_carlo, freqs, sampling_freq, filter_order, weight_vector, verbose, inv
-):
-    LSFIR(
-        H=monte_carlo["H"],
-        N=filter_order,
-        f=freqs,
-        Fs=sampling_freq,
-        tau=filter_order // 2,
-        weights=weight_vector,
-        verbose=verbose,
-        inv=inv,
-        UH=monte_carlo["UH"],
-        mc_runs=2,
-    )
-    _print_during_test_to_avoid_timeout(capsys)
 
 
 @given(hypothesis_dimension(min_value=4, max_value=8))
@@ -803,51 +441,6 @@ def test_LSFIR_with_wrong_type_UH(monte_carlo, freqs, sampling_freq, filter_orde
             UH=cast(np.ndarray, uh_list),
             mc_runs=2,
         )
-
-
-@given(hypothesis_dimension(min_value=4, max_value=8))
-@settings(
-    deadline=None,
-    suppress_health_check=[
-        *settings.default.suppress_health_check,
-        HealthCheck.too_slow,
-        HealthCheck.function_scoped_fixture,
-    ],
-)
-@pytest.mark.slow
-def test_compare_different_dtypes_LSFIR(
-    capsys,
-    monte_carlo,
-    complex_H_with_UH,
-    freqs,
-    sampling_freq,
-    filter_order,
-):
-    b_real_imaginary, ub_real_imaginary = LSFIR(
-        H=monte_carlo["H"],
-        N=filter_order,
-        f=freqs,
-        Fs=sampling_freq,
-        tau=filter_order // 2,
-        inv=True,
-        verbose=True,
-        UH=monte_carlo["UH"],
-        mc_runs=10000,
-    )
-    b_complex, ub_complex = LSFIR(
-        H=complex_H_with_UH["H"],
-        N=filter_order,
-        f=freqs,
-        Fs=sampling_freq,
-        tau=filter_order // 2,
-        inv=True,
-        verbose=True,
-        UH=monte_carlo["UH"],
-        mc_runs=10000,
-    )
-    _print_during_test_to_avoid_timeout(capsys)
-    assert_allclose(b_real_imaginary, b_complex, rtol=4e-2)
-    assert_allclose(ub_real_imaginary, ub_complex, rtol=6e-1)
 
 
 @given(hypothesis_dimension(min_value=4, max_value=8))
@@ -1057,117 +650,6 @@ def test_too_small_number_of_monte_carlo_runs_LSFIR(
             inv=inv,
             mc_runs=1,
         )
-
-
-@given(hypothesis_dimension(min_value=4, max_value=8))
-@settings(
-    deadline=None,
-    suppress_health_check=[
-        *settings.default.suppress_health_check,
-        HealthCheck.too_slow,
-        HealthCheck.function_scoped_fixture,
-    ],
-    max_examples=10,
-)
-@pytest.mark.slow
-def test_compare_LSFIR_with_svd_and_with_mc(
-    capsys, monte_carlo, freqs, sampling_freq, filter_order
-):
-    b_fir_svd, Ub_fir_svd = LSFIR(
-        H=monte_carlo["H"],
-        N=filter_order,
-        f=freqs,
-        Fs=sampling_freq,
-        tau=filter_order // 2,
-        verbose=True,
-        inv=True,
-        UH=monte_carlo["UH"],
-    )
-    b_fir_mc, Ub_fir_mc = LSFIR(
-        H=monte_carlo["H"],
-        N=filter_order,
-        f=freqs,
-        Fs=sampling_freq,
-        tau=filter_order // 2,
-        verbose=True,
-        inv=True,
-        UH=monte_carlo["UH"],
-        mc_runs=10000,
-    )
-    _print_during_test_to_avoid_timeout(capsys)
-    assert_allclose(b_fir_mc, b_fir_svd, rtol=9e-2)
-    assert_allclose(Ub_fir_mc, Ub_fir_svd, atol=6e-1, rtol=6e-1)
-
-
-@given(hypothesis_dimension(min_value=4, max_value=8))
-@settings(
-    deadline=None,
-    suppress_health_check=[
-        *settings.default.suppress_health_check,
-        HealthCheck.too_slow,
-        HealthCheck.function_scoped_fixture,
-    ],
-)
-@pytest.mark.slow
-def test_compare_invLSFIR_uncMC_LSFIR(
-    capsys, monte_carlo, freqs, sampling_freq, filter_order
-):
-    b_fir_mc, Ub_fir_mc = invLSFIR_uncMC(
-        H=monte_carlo["H"],
-        UH=monte_carlo["UH"],
-        N=filter_order,
-        tau=filter_order // 2,
-        f=freqs,
-        Fs=sampling_freq,
-    )
-    b_fir, Ub_fir = LSFIR(
-        H=monte_carlo["H"],
-        N=filter_order,
-        f=freqs,
-        Fs=sampling_freq,
-        tau=filter_order // 2,
-        inv=True,
-        UH=monte_carlo["UH"],
-        mc_runs=10000,
-    )
-    _print_during_test_to_avoid_timeout(capsys)
-    assert_allclose(b_fir_mc, b_fir, rtol=4e-2)
-    assert_allclose(Ub_fir_mc, Ub_fir, atol=6e-1, rtol=6e-1)
-
-
-@given(hypothesis_dimension(min_value=4, max_value=8))
-@settings(
-    deadline=None,
-    suppress_health_check=[
-        *settings.default.suppress_health_check,
-        HealthCheck.too_slow,
-        HealthCheck.function_scoped_fixture,
-    ],
-)
-@pytest.mark.slow
-def test_compare_invLSFIR_unc_LSFIR_only_by_filter_coefficients(
-    capsys, monte_carlo, freqs, sampling_freq, filter_order
-):
-    b_fir_mc, Ub_fir_mc = invLSFIR_unc(
-        H=monte_carlo["H"],
-        UH=monte_carlo["UH"],
-        N=filter_order,
-        tau=filter_order // 2,
-        f=freqs,
-        Fs=sampling_freq,
-    )
-    b_fir, Ub_fir = LSFIR(
-        H=monte_carlo["H"],
-        N=filter_order,
-        f=freqs,
-        Fs=sampling_freq,
-        tau=filter_order // 2,
-        inv=True,
-        UH=monte_carlo["UH"],
-    )
-    _print_during_test_to_avoid_timeout(capsys)
-    assert_allclose(b_fir_mc, b_fir)
-    assert_allclose(Ub_fir_mc, Ub_fir, atol=6e-1, rtol=6e-1)
 
 
 @given(hypothesis_dimension(min_value=4, max_value=8))
