+++ test/test_lsfir/test_compare_ls_fir_with_svd_and_with_mc.py
@@ -1,0 +1,51 @@
+import pytest
+from hypothesis import given, HealthCheck, settings
+from numpy.testing import assert_allclose
+
+# noinspection PyProtectedMember
+from PyDynamic.model_estimation.fit_filter import (
+    LSFIR,
+)
+from ..conftest import (
+    _print_during_test_to_avoid_timeout,
+    hypothesis_dimension,
+)
+
+
+@given(hypothesis_dimension(min_value=4, max_value=8))
+@settings(
+    deadline=None,
+    suppress_health_check=[
+        *settings.default.suppress_health_check,
+        HealthCheck.too_slow,
+        HealthCheck.function_scoped_fixture,
+    ],
+)
+@pytest.mark.slow
+def test_compare_LSFIR_with_svd_and_with_mc(
+    capsys, monte_carlo, freqs, sampling_freq, filter_order
+):
+    b_fir_svd, Ub_fir_svd = LSFIR(
+        H=monte_carlo["H"],
+        N=filter_order,
+        f=freqs,
+        Fs=sampling_freq,
+        tau=filter_order // 2,
+        verbose=True,
+        inv=True,
+        UH=monte_carlo["UH"],
+    )
+    b_fir_mc, Ub_fir_mc = LSFIR(
+        H=monte_carlo["H"],
+        N=filter_order,
+        f=freqs,
+        Fs=sampling_freq,
+        tau=filter_order // 2,
+        verbose=True,
+        inv=True,
+        UH=monte_carlo["UH"],
+        mc_runs=10000,
+    )
+    _print_during_test_to_avoid_timeout(capsys)
+    assert_allclose(b_fir_mc, b_fir_svd, rtol=9e-2)
+    assert_allclose(Ub_fir_mc, Ub_fir_svd, atol=6e-1, rtol=6e-1)
